+++ index.js
@@ -13,7 +13,6 @@ import BugReporting from './modules/BugReporting';
 import APM from './modules/APM';
 import Surveys from './modules/Surveys';
 import FeatureRequests from './modules/FeatureRequests';
-import Chats from './modules/Chats';
 import Replies from './modules/Replies';
 import CrashReporting from './modules/CrashReporting';
 import NetworkLogger from './modules/NetworkLogger';
@@ -31,36 +30,18 @@ const firstScreen = "Initial Screen";
  */
 const InstabugModule = {
 
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Instabug.start}
-   * Starts the SDK.
-   * This is the main SDK method that does all the magic. This is the only
-   * method that SHOULD be called.
-   * Should be called in constructor of the app registery component
-   * @param {string} token The token that identifies the app, you can find
-   * it on your dashboard.
-   * @param {invocationEvent} invocationEvent The event that invokes
-   * the SDK's UI.
-   */
-  startWithToken(token, invocationEvent) {
-    this.start(token, invocationEvent);
-  },
-
   /**
    * Starts the SDK.
    * This is the main SDK method that does all the magic. This is the only
    * method that SHOULD be called.
-   * Should be called in constructor of the app registery component
+   * Should be called in constructor of the AppRegistry component
    * @param {string} token The token that identifies the app, you can find
    * it on your dashboard.
    * @param {invocationEvent} invocationEvent The event that invokes
    * the SDK's UI.
    */
   start: function (token, invocationEvent) {
-    if (Platform.OS === 'ios') {
-      Instabug.startWithToken(token, invocationEvent);
-    }
+    Instabug.start(token, invocationEvent);
     _isFirstScreen = true;
     _currentScreen = firstScreen;
     setTimeout(function () {
@@ -80,30 +61,6 @@ const InstabugModule = {
    */
   setUserData(userData) {
     Instabug.setUserData(userData);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link BugReporting.setAutoScreenRecordingEnabled}
-   * Enable/Disable screen recording
-   * @param {boolean} autoScreenRecordingEnabled boolean for enable/disable
-   * screen recording on crash feature
-   */
-  setAutoScreenRecordingEnabled(autoScreenRecordingEnabled) {
-    Instabug.setAutoScreenRecordingEnabled(autoScreenRecordingEnabled);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link BugReporting.setAutoScreenRecordingMaxDuration}
-   * Sets auto screen recording maximum duration
-   *
-   * @param autoScreenRecordingMaxDuration maximum duration of the screen recording video
-   *                                       in seconds
-   * The maximum duration is 30 seconds
-   */
-  setAutoScreenRecordingMaxDuration(autoScreenRecordingMaxDuration) {
-    Instabug.setAutoScreenRecordingMaxDuration(autoScreenRecordingMaxDuration);
   },
 
   /**
@@ -128,27 +85,6 @@ const InstabugModule = {
       Instabug.setIBGLogPrintsToConsole(printsToConsole);
   },
 
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link CrashReporting.setEnabled}
-   * Report un-caught exceptions to Instabug dashboard
-   * We don't send exceptions from __DEV__, since it's way too noisy!
-   */
-  setCrashReportingEnabled(enableCrashReporter) {
-    Instabug.setCrashReportingEnabled(enableCrashReporter);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link BugReporting.setDidSelectPromptOptionHandler}
-   * Sets a block of code to be executed when a prompt option is selected.
-   * @param {function} didSelectPromptOptionHandler - A block of code that
-   *                  gets executed when a prompt option is selected.
-   */
-  setDidSelectPromptOptionHandler(didSelectPromptOptionHandler) {
-    BugReporting.setDidSelectPromptOptionHandler(didSelectPromptOptionHandler);
-  },
-
   /**
    * The session profiler is enabled by default and it attaches to the bug and
    * crash reports the following information during the last 60 seconds before the report is sent.
@@ -169,62 +105,6 @@ const InstabugModule = {
     if (Platform.OS === 'ios') {
       Instabug.setSdkDebugLogsLevel(sdkDebugLogsLevel);
     }
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Replies.getUnreadRepliesCount}
-   * Returns the number of unread messages the user currently has.
-   * Use this method to get the number of unread messages the user
-   * has, then possibly notify them about it with your own UI.
-   * @param {messageCountCallback} messageCountCallback callback with argument
-   * Notifications count, or -1 in case the SDK has not been initialized.
-   */
-
-  getUnreadMessagesCount(messageCountCallback) {
-    Instabug.getUnreadMessagesCount(messageCountCallback);
-  },
-
-  /**
-   * @deprecated use {@link Replies.setPushNotificationsEnabled}
-   * Enables/disables the use of push notifications in the SDK.
-   * Defaults to YES.
-   * @param {boolean} isPushNotificationEnabled A boolean to indicate whether push
-   * notifications are enabled or disabled.
-   */
-  setPushNotificationsEnabled(isPushNotificationEnabled) {
-    Replies.setPushNotificationsEnabled(isPushNotificationEnabled);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link BugReporting.setInvocationOptions}
-   * Sets whether users are required to enter an email address or not when
-   * sending reports.
-   * Defaults to YES.
-   * @param {boolean} isEmailFieldRequired A boolean to indicate whether email
-   * field is required or not.
-   * @param {actionTypes} actionTypes An enum that indicates which action
-   *                                  types will have the isEmailFieldRequired
-   */
-
-  setEmailFieldRequiredForActions(isEmailFieldRequired, actionTypes) {
-    Instabug.setEmailFieldRequiredForActions(isEmailFieldRequired, actionTypes);
-  },
-
-  /**
-   * @deprecated use {@link BugReporting.setFloatingButtonEdge}
-   * Sets the default edge and offset from the top at which the floating button
-   * will be shown. Different orientations are already handled.
-   * Default for `floatingButtonEdge` is `rectEdge.maxX`.
-   * Default for `floatingButtonOffsetFromTop` is 50
-   * @param {rectEdge} floatingButtonEdge `maxX` to show on the right,
-   * or `minX` to show on the left.
-   * @param {number} offsetFromTop floatingButtonOffsetFromTop Top offset for
-   * floating button.
-   */
-  setFloatingButtonEdge(floatingButtonEdge, offsetFromTop) {
-    BugReporting.setFloatingButtonEdge(floatingButtonEdge, offsetFromTop);
   },
 
   /**
@@ -281,18 +161,6 @@ const InstabugModule = {
     Instabug.getTags(tagsCallback);
   },
 
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Instabug.setString}
-   * Overrides any of the strings shown in the SDK with custom ones.
-   * Allows you to customize any of the strings shown to users in the SDK.
-   * @param {string} string String value to override the default one.
-   * @param {strings} key Key of string to override.
-   */
-  setStringToKey(string, key) {
-    this.setString(key, string);
-  },
-
   /**
    * Overrides any of the strings shown in the SDK with custom ones.
    * Allows you to customize any of the strings shown to users in the SDK.
@@ -304,45 +172,6 @@ const InstabugModule = {
   },
 
   /**
-   * @deprecated use {@link BugReporting.setEnabledAttachmentTypes}
-   * Sets whether attachments in bug reporting and in-app messaging are enabled or not.
-   * @param {boolean} screenshot A boolean to enable or disable screenshot attachments.
-   * @param {boolean} extraScreenshot A boolean to enable or disable extra
-   * screenshot attachments.
-   * @param {boolean} galleryImage A boolean to enable or disable gallery image
-   * attachments. In iOS 10+,NSPhotoLibraryUsageDescription should be set in
-   * info.plist to enable gallery image attachments.
-   * @param {boolean} screenRecording A boolean to enable or disable screen recording attachments.
-   */
-  setEnabledAttachmentTypes(
-    screenshot,
-    extraScreenshot,
-    galleryImage,
-    screenRecording
-  ) {
-    BugReporting.setEnabledAttachmentTypes(
-      screenshot,
-      extraScreenshot,
-      galleryImage,
-      screenRecording
-    );
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Instabug.identifyUser}
-   * Sets the default value of the user's email and hides the email field from the reporting UI
-   * and set the user's name to be included with all reports.
-   * It also reset the chats on device to that email and removes user attributes,
-   * user data and completed surveys.
-   * @param {string} email Email address to be set as the user's email.
-   * @param {string} name Name of the user to be set.
-   */
-  identifyUserWithEmail(email, name) {
-    this.identifyUser(email, name);
-  },
-
-  /**
    * Sets the default value of the user's email and hides the email field from the reporting UI
    * and set the user's name to be included with all reports.
    * It also reset the chats on device to that email and removes user attributes,
@@ -351,7 +180,7 @@ const InstabugModule = {
    * @param {string} name Name of the user to be set.
    */
   identifyUser(email, name) {
-    Instabug.identifyUserWithEmail(email, name);
+    Instabug.identifyUser(email, name);
   },
 
   /**
@@ -363,16 +192,6 @@ const InstabugModule = {
     Instabug.logOut();
   },
 
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Instabug.logUserEvent}
-   * Logs a user event that happens through the lifecycle of the application.
-   * Logged user events are going to be sent with each report, as well as at the end of a session.
-   * @param {string} name Event name.
-   */
-  logUserEventWithName(name) {
-    this.logUserEvent(name);
-  },
 
   /**
    * Logs a user event that happens through the lifecycle of the application.
@@ -380,7 +199,7 @@ const InstabugModule = {
    * @param {string} name Event name.
    */
   logUserEvent(name) {
-    Instabug.logUserEventWithName(name);
+    Instabug.logUserEvent(name);
   },
 
   /**
@@ -547,57 +366,6 @@ const InstabugModule = {
     Instabug.clearAllUserAttributes();
   },
 
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Replies.setInAppNotificationsEnabled}
-   * Enables/disables showing in-app notifications when the user receives a
-   * new message.
-   * @param {boolean} isChatNotificationEnabled A boolean to set whether
-   * notifications are enabled or disabled.
-   */
-
-  setChatNotificationEnabled(isChatNotificationEnabled) {
-    Instabug.setChatNotificationEnabled(isChatNotificationEnabled);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Replies.setOnNewReplyReceivedCallback}
-   * Sets a block of code that gets executed when a new message is received.
-   * @param {function} onNewMessageHandler - A callback that gets
-   * executed when a new message is received.
-   */
-  setOnNewMessageHandler(onNewMessageHandler) {
-    Replies.setOnNewReplyReceivedHandler(onNewMessageHandler);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link BugReporting.setViewHierarchyEnabled}
-   * @summary Enables/disables inspect view hierarchy when reporting a bug/feedback.
-   * @param {boolean} viewHierarchyEnabled A boolean to set whether view hierarchy are enabled
-   * or disabled.
-   */
-  setViewHierarchyEnabled(viewHierarchyEnabled) {
-    Instabug.setViewHierarchyEnabled(viewHierarchyEnabled);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Surveys.setEnabled}
-   * @summary Sets whether surveys are enabled or not.
-   * If you disable surveys on the SDK but still have active surveys on your Instabug dashboard,
-   * those surveys are still going to be sent to the device, but are not going to be
-   * shown automatically.
-   * To manually display any available surveys, call `Instabug.showSurveyIfAvailable()`.
-   * Defaults to `true`.
-   * @param {boolean} surveysEnabled A boolean to set whether Instabug Surveys is enabled or disabled.
-   */
-
-  setSurveysEnabled(surveysEnabled) {
-    Instabug.setSurveysEnabled(surveysEnabled);
-  },
-
   /**
    * Enable/Disable debug logs from Instabug SDK
    * Default state: disabled
@@ -630,34 +398,6 @@ const InstabugModule = {
     }
   },
 
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Replies.setInAppNotificationSound}
-   * Set whether new in app notification received will play a small sound notification
-   * or not (Default is {@code false})
-   *
-   * @param shouldPlaySound desired state of conversation sounds
-   * @since 4.1.0
-   */
-
-  setEnableInAppNotificationSound(shouldPlaySound) {
-    if (Platform.OS === 'android') {
-      Instabug.setEnableInAppNotificationSound(shouldPlaySound);
-    }
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link CrashReporting.reportJSException}
-   * Send handled JS error object
-   *
-   * @param errorObject Error object to be sent to Instabug's servers
-   */
-
-  reportJSException(errorObject) {
-    CrashReporting.reportJSException(errorObject);
-  },
-
   /**
    * @summary Checks whether app is development/Beta testing OR live
    * Note: This API is iOS only
@@ -669,33 +409,6 @@ const InstabugModule = {
     if (Platform.OS === 'ios') {
       Instabug.isRunningLive(runningLiveCallBack);
     }
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link BugReporting.setVideoRecordingFloatingButtonPosition}
-   * Sets the default position at which the Instabug screen recording button will be shown.
-   * Different orientations are already handled.
-   * (Default for `position` is `bottomRight`)
-   *
-   * @param position is of type IBGPosition `topLeft` to show on the top left of screen,
-   * or `bottomRight` to show on the bottom right of scrren.
-   */
-  setVideoRecordingFloatingButtonPosition(position) {
-    BugReporting.setVideoRecordingFloatingButtonPosition(position);
-  },
-
-  /* istanbul ignore next */
-  /**
-   * @deprecated use {@link Surveys.setShouldShowWelcomeScreen}
-   * Setting an option for all the surveys to show a welcome screen before
-   * the user starts taking the survey.
-   * @param shouldShowWelcomeScreen A boolean for setting whether the
-   *                                welcome screen should show.
-   *
-   */
-  setShouldShowSurveysWelcomeScreen(shouldShowWelcomeScreen) {
-    Instabug.setShouldShowSurveysWelcomeScreen(shouldShowWelcomeScreen);
   },
 
   /**
@@ -732,17 +445,34 @@ const InstabugModule = {
   },
 
   /**
+   * @deprecated Use {@link Instabug.addPrivateView} instead. 
+   * 
    * Hides component from screenshots, screen recordings and view hierarchy.
    * @param {Object} viewRef the ref of the component to hide
    */
   setPrivateView(viewRef) {
-    const nativeTag = findNodeHandle(viewRef);
-    if (Platform.OS === 'ios') {
-      Instabug.hideView(nativeTag);
-    } else {
-      Instabug.hideView([nativeTag]);
-    }
+    this.addPrivateView(viewRef);
   },
+
+  /**
+   * Hides component from screenshots, screen recordings and view hierarchy.
+   * @param {Object} viewRef the ref of the component to hide
+   */
+  addPrivateView(viewRef) {
+    const nativeTag = findNodeHandle(viewRef);
+    Instabug.addPrivateView(nativeTag);
+  },
+
+  /**
+   * Removes component from the set of hidden views. The component will show again in 
+   * screenshots, screen recordings and view hierarchy.
+   * @param {Object} viewRef the ref of the component to remove from hidden views
+   */
+  removePrivateView(viewRef) {
+    const nativeTag = findNodeHandle(viewRef);
+    Instabug.removePrivateView(nativeTag);
+  },
+
   /**
    * Shows default Instabug prompt.
    */
@@ -907,18 +637,6 @@ const InstabugModule = {
   },
 
   /**
-   *  The options used upon invocating the SDK
-   * @readonly
-   * @enum {number}
-   */
-  invocationOptions: {
-    invocationOptionsEmailFieldHidden: Instabug.emailFieldHidden,
-    invocationOptionsEmailFieldOptional: Instabug.emailFieldOptional,
-    invocationOptionsCommentFieldRequired: Instabug.commentFieldRequired,
-    invocationOptionsDisablePostSendingDialog: Instabug.disablePostSendingDialog
-  },
-
-  /**
    * Verbosity level of the SDK debug logs. This has nothing to do with IBGLog,
    * and only affect the logs used to debug the SDK itself.
    * @readonly
@@ -1039,7 +757,6 @@ const InstabugModule = {
     invalidCommentMessage: Instabug.invalidCommentMessage,
     invalidCommentTitle: Instabug.invalidCommentTitle,
     invocationHeader: Instabug.invocationHeader,
-    startChats: Instabug.startChats,
     reportQuestion: Instabug.reportQuestion,
     reportBug: Instabug.reportBug,
     reportFeedback: Instabug.reportFeedback,
@@ -1047,6 +764,7 @@ const InstabugModule = {
     commentFieldHintForBugReport: Instabug.commentFieldHintForBugReport,
     commentFieldHintForFeedback: Instabug.commentFieldHintForFeedback,
     commentFieldHintForQuestion: Instabug.commentFieldHintForQuestion,
+    videoPressRecord: Instabug.videoPressRecord,
     addVideoMessage: Instabug.addVideoMessage,
     addVoiceMessage: Instabug.addVoiceMessage,
     addImageFromGallery: Instabug.addImageFromGallery,
@@ -1061,7 +779,6 @@ const InstabugModule = {
     recordingMessageToReleaseText: Instabug.recordingMessageToReleaseText,
     conversationsHeaderTitle: Instabug.conversationsHeaderTitle,
     screenshotHeaderTitle: Instabug.screenshotHeaderTitle,
-    chatsNoConversationsHeadlineText: Instabug.chatsNoConversationsHeadlineText,
     doneButtonText: Instabug.doneButtonText,
     okButtonText: Instabug.okButtonTitle,
     cancelButtonText: Instabug.cancelButtonTitle,
@@ -1069,10 +786,6 @@ const InstabugModule = {
     audio: Instabug.audio,
     video: Instabug.video,
     image: Instabug.image,
-    /**
-     * @deprecated use {@link Instabug.strings.conversationsHeaderTitle}
-     */
-    chatsHeaderTitle: Instabug.chatsHeaderTitle,
     team: Instabug.team,
     messagesNotification: Instabug.messagesNotification,
     messagesNotificationAndOthers: Instabug.messagesNotificationAndOthers,
@@ -1095,8 +808,6 @@ const InstabugModule = {
       Instabug.welcomeMessageLiveWelcomeStepTitle,
     welcomeMessageLiveWelcomeStepContent:
       Instabug.welcomeMessageLiveWelcomeStepContent,
-    surveysCustomThanksTitle: Instabug.surveysCustomThanksTitle,
-    surveysCustomThanksSubTitle: Instabug.surveysCustomThanksSubTitle,
     surveysStoreRatingThanksTitle: Instabug.surveysStoreRatingThanksTitle,
     surveysStoreRatingThanksSubtitle: Instabug.surveysStoreRatingThanksSubtitle,
     reportBugDescription: Instabug.reportBugDescription,
@@ -1107,7 +818,14 @@ const InstabugModule = {
     discardAlertMessage: Instabug.discardAlertMessage,
     discardAlertCancel: Instabug.discardAlertCancel,
     discardAlertAction: Instabug.discardAlertAction,
-    addAttachmentButtonTitleStringName: Instabug.addAttachmentButtonTitleStringName
+    addAttachmentButtonTitleStringName: Instabug.addAttachmentButtonTitleStringName,
+    reportReproStepsDisclaimerBody: Instabug.reportReproStepsDisclaimerBody,
+    reportReproStepsDisclaimerLink: Instabug.reportReproStepsDisclaimerLink,
+    reproStepsProgressDialogBody: Instabug.reproStepsProgressDialogBody,
+    reproStepsListHeader: Instabug.reproStepsListHeader,
+    reproStepsListDescription: Instabug.reproStepsListDescription,
+    reproStepsListEmptyStateDescription: Instabug.reproStepsListEmptyStateDescription,
+    reproStepsListItemTitle: Instabug.reproStepsListItemTitle,
   },
 
 };
@@ -1116,7 +834,6 @@ export {
   BugReporting,
   Surveys,
   FeatureRequests,
-  Chats,
   Replies,
   CrashReporting,
   NetworkLogger,