+++ ios/RNInstabug/InstabugReactBridge.m
@@ -38,7 +38,7 @@ RCT_EXPORT_MODULE(Instabug)
     return dispatch_get_main_queue();
 }
 
-RCT_EXPORT_METHOD(startWithToken:(NSString *)token invocationEvents:(NSArray*)invocationEventsArray) {
+RCT_EXPORT_METHOD(start:(NSString *)token invocationEvents:(NSArray*)invocationEventsArray) {
      SEL setPrivateApiSEL = NSSelectorFromString(@"setCurrentPlatform:");
     if ([[Instabug class] respondsToSelector:setPrivateApiSEL]) {
         NSInteger *platform = IBGPlatformReactNative;
@@ -243,7 +243,7 @@ RCT_EXPORT_METHOD(clearFileAttachments) {
     [Instabug clearFileAttachments];
 }
 
-RCT_EXPORT_METHOD(identifyUserWithEmail:(NSString *)email name:(NSString *)name) {
+RCT_EXPORT_METHOD(identifyUser:(NSString *)email name:(NSString *)name) {
     [Instabug identifyUserWithEmail:email name:name];
 }
 
@@ -277,11 +277,7 @@ RCT_EXPORT_METHOD(clearAllUserAttributes) {
     }
 }
 
-RCT_EXPORT_METHOD(setViewHierarchyEnabled:(BOOL)viewHierarchyEnabled) {
-    Instabug.shouldCaptureViewHierarchy = viewHierarchyEnabled;
-}
-
-RCT_EXPORT_METHOD(logUserEventWithName:(NSString *)name) {
+RCT_EXPORT_METHOD(logUserEvent:(NSString *)name) {
     [Instabug logUserEventWithName:name];
 }
 
@@ -413,10 +409,14 @@ RCT_EXPORT_METHOD(networkLog:(NSDictionary *) networkData) {
     }
 }
 
-RCT_EXPORT_METHOD(hideView: (nonnull NSNumber *)reactTag) {
+RCT_EXPORT_METHOD(addPrivateView: (nonnull NSNumber *)reactTag) {
     UIView* view = [self.bridge.uiManager viewForReactTag:reactTag];
     view.instabug_privateView = true;
-    
+}
+
+RCT_EXPORT_METHOD(removePrivateView: (nonnull NSNumber *)reactTag) {
+    UIView* view = [self.bridge.uiManager viewForReactTag:reactTag];
+    view.instabug_privateView = false;
 }
 
 RCT_EXPORT_METHOD(show) {
@@ -463,7 +463,7 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               
               @"dismissTypeSubmit": @(IBGDismissTypeSubmit),
               @"dismissTypeCancel": @(IBGDismissTypeCancel),
-              @"dismissTypeAddAtttachment": @(IBGDismissTypeAddAttachment),
+              @"dismissTypeAddAttachment": @(IBGDismissTypeAddAttachment),
               
               @"reproStepsEnabled": @(IBGUserStepsModeEnable),
               @"reproStepsDisabled": @(IBGUserStepsModeDisable),
@@ -525,12 +525,6 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               @"sdkDebugLogsLevelError": @(IBGSDKDebugLogsLevelError),
               @"sdkDebugLogsLevelNone": @(IBGSDKDebugLogsLevelNone),
 
-
-              @"emailFieldHidden": @(IBGBugReportingInvocationOptionEmailFieldHidden),
-              @"emailFieldOptional": @(IBGBugReportingInvocationOptionEmailFieldOptional),
-              @"commentFieldRequired": @(IBGBugReportingInvocationOptionCommentFieldRequired),
-              @"disablePostSendingDialog": @(IBGBugReportingInvocationOptionDisablePostSendingDialog),
-              
               @"colorThemeLight": @(IBGColorThemeLight),
               @"colorThemeDark": @(IBGColorThemeDark),
               
@@ -552,7 +546,6 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               @"invalidCommentTitle": kIBGInvalidCommentTitleStringName,
               @"invocationHeader": kIBGInvocationTitleStringName,
               //@"talkToUs": kIBGTalkToUsStringName,
-              @"startChats": kIBGAskAQuestionStringName,
               @"reportQuestion": kIBGAskAQuestionStringName,
               @"reportBug": kIBGReportBugStringName,
               @"reportFeedback": kIBGReportFeedbackStringName,
@@ -568,7 +561,6 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               @"audioRecordingPermissionDeniedMessage": kIBGAudioRecordingPermissionDeniedMessageStringName,
               @"microphonePermissionAlertSettingsButtonTitle": kIBGMicrophonePermissionAlertSettingsButtonTitleStringName,
               @"conversationsHeaderTitle": kIBGChatsTitleStringName,
-              @"chatsHeaderTitle": kIBGChatsTitleStringName,
               @"team": kIBGTeamStringName,
               @"recordingMessageToHoldText": kIBGRecordingMessageToHoldTextStringName,
               @"recordingMessageToReleaseText": kIBGRecordingMessageToReleaseTextStringName,
@@ -582,7 +574,7 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               @"screenRecording": kIBGScreenRecordingStringName,
               @"image": kIBGImageStringName,
               @"surveyEnterYourAnswer": kIBGSurveyEnterYourAnswerTextPlaceholder,
-              @"videPressRecord": kIBGVideoPressRecordTitle,
+              @"videoPressRecord": kIBGVideoPressRecordTitle,
               @"collectingDataText": kIBGCollectingDataText,
               @"thankYouAlertText": kIBGThankYouAlertMessageStringName,
               
@@ -594,9 +586,6 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               @"welcomeMessageBetaFinishStepContent": kIBGBetaWelcomeMessageFinishStepContent,
               @"welcomeMessageLiveWelcomeStepTitle": kIBGLiveWelcomeMessageTitle,
               @"welcomeMessageLiveWelcomeStepContent": kIBGLiveWelcomeMessageContent,
-              
-              @"surveysCustomThanksTitle": kIBGCustomSurveyThankYouTitleText,
-              @"surveysCustomThanksSubtitle": kIBGCustomSurveyThankYouDescriptionText,
               
               @"surveysStoreRatingThanksTitle": kIBGStoreRatingThankYouTitleText,
               @"surveysStoreRatingThanksSubtitle": kIBGStoreRatingThankYouDescriptionText,
@@ -610,7 +599,14 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
               @"discardAlertMessage": kIBGDiscardAlertMessage,
               @"discardAlertCancel": kIBGDiscardAlertCancel,
               @"discardAlertAction": kIBGDiscardAlertAction,
-              @"addAttachmentButtonTitleStringName": kIBGAddAttachmentButtonTitleStringName
+              @"addAttachmentButtonTitleStringName": kIBGAddAttachmentButtonTitleStringName,
+              @"reportReproStepsDisclaimerBody": kIBGReproStepsDisclaimerBody,
+              @"reportReproStepsDisclaimerLink": kIBGReproStepsDisclaimerLink,
+              @"reproStepsProgressDialogBody": kIBGProgressViewTitle,
+              @"reproStepsListHeader": kIBGReproStepsListTitle,
+              @"reproStepsListDescription": kIBGReproStepsListHeader,
+              @"reproStepsListEmptyStateDescription": kIBGReproStepsListEmptyStateLabel,
+              @"reproStepsListItemTitle": kIBGReproStepsListItemName
             };
 };
 
@@ -658,7 +654,7 @@ RCTLogFunction InstabugReactLogFunction = ^(
     
     switch(level) {
         case RCTLogLevelTrace:
-            RNIBGLog(IBGLogLevelTrace, formatString, log);
+            RNIBGLog(IBGLogLevelVerbose, formatString, log);
             break;
         case RCTLogLevelInfo:
             RNIBGLog(IBGLogLevelInfo, formatString, log);
@@ -670,7 +666,7 @@ RCTLogFunction InstabugReactLogFunction = ^(
             RNIBGLog(IBGLogLevelError, formatString, log);
             break;
         case RCTLogLevelFatal:
-            RNIBGLog(IBGLogLevelFatal, formatString, log);
+            RNIBGLog(IBGLogLevelError, formatString, log);
             break;
     }
 };