+++ jest/mockChats.js
@@ -1,8 +1,0 @@
-jest.mock('NativeModules', () => {
-    return {
-      IBGChats: {
-        setEnabled: jest.fn(),
-        show: jest.fn()
-      },
-    };
-  });